Massive send modal files renaming
