+++ src/modals/SendModals/Script/CheckTxModalContent.tsx
@@ -16,8 +16,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import InfoBox from '../../components/InfoBox'
-import { DeployContractTxData } from './BuildDeployContractTx'
+import { ScriptTxData } from '../../../types/transactions'
 import {
   AlphAmountInfo,
   BytecodeInfo,
@@ -27,19 +26,18 @@ import {
   FeeInfo,
   FromAddressInfo,
   ModalContent
-} from './utils'
+} from '../utils'
 
-const CheckDeployContractTx = ({ data, fees, onSend, onCancel }: CheckTxProps<DeployContractTxData>) => (
+const CheckScriptTx = ({ data, fees, onSend, onCancel }: CheckTxProps<ScriptTxData>) => (
   <>
     <ModalContent>
       <FromAddressInfo fromAddress={data.fromAddress} />
       <BytecodeInfo bytecode={data.bytecode} />
       <AlphAmountInfo expectedAmount={expectedAmount(data, fees)} />
-      {data.issueTokenAmount && <InfoBox text={data.issueTokenAmount} label="Issue token amount" wordBreak />}
       <FeeInfo fees={fees} />
     </ModalContent>
     <CheckTxFooter onSend={onSend} onCancel={onCancel} />
   </>
 )
 
-export default CheckDeployContractTx
+export default CheckScriptTx
